+++ src/components/Authentication/index.js
@@ -1,24 +1,15 @@
-import React from 'react';
+import React, { lazy } from 'react';
 
-import { Box, Grommet } from 'grommet';
-import { grommet } from 'grommet/themes';
-import Routes from './routes';
+import { Route, Switch } from 'react-router-dom';
 
-export default function(props) {
+const Login = lazy(() => import('./Login'));
+const UpdatePassword = lazy(() => import('./UpdatePassword'));
+
+export default function({ match: { path }, login, loading, error }) {
   return (
-    <Grommet full theme={grommet}>
-      <Box fill align="center" justify="center">
-        <Box
-          style={{
-            borderRadius: '10px',
-            boxShadow: '0px 0px 20px 5px rgba(0, 0, 0, 0.1)'
-          }}
-          width="medium"
-          pad="medium"
-        >
-          <Routes {...props} />
-        </Box>
-      </Box>
-    </Grommet>
+    <Switch>
+      <Route exact path={`/login`} component={Login} />
+      <Route exact path={`/update-password`} component={UpdatePassword} />
+    </Switch>
   );
 }
