+++ src/reducers.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import user from './compositions/Authentication/user.reducer';
-import authentication from './compositions/Authentication/authentication.reducer';
+import user from './components/Authentication/user.reducer';
+import authentication from './components/Authentication/authentication.reducer';
 
 export default combineReducers({ authentication, user });
